Split SDLFrame to ease integration of ImGui.
